Change server DAO
Ticket: #114
