+++ src/app/application.types.ts
@@ -4,12 +4,12 @@ export const APPLICATION_TYPES = {
     FilterService: Symbol.for('FilterService'),
     GroupService: Symbol.for('GroupService'),
     IsolateService: Symbol.for('IsolateService'),
+    IsolateViewGateway: Symbol.for('IsolateViewGateway'),
     IsolateGateway: Symbol.for('IsolateGateway'),
     MicroorganismGateway: Symbol.for('MicroorganismGateway'),
     SamplingContextGateway: Symbol.for('SamplingContextGateway'),
     SamplingStageGateway: Symbol.for('SamplingStageGateway'),
     MatrixGateway: Symbol.for('MatrixGateway'),
-    ProgramGateway: Symbol.for('ProgramGateway'),
     OriginGateway: Symbol.for('OriginGateway'),
     FederalStateGateway: Symbol.for('FederalStateGateway'),
     CategoryGateway: Symbol.for('CategoryGateway'),
