--- src/app/query/model/program.model.ts
@@ -1,7 +1,0 @@
-import { EntityGateway } from './shared.model';
-
-export interface Program {
-    samplingYear: string;
-}
-
-export interface ProgramGateway extends EntityGateway<Program> {}
